yokai_inventory_scraper/requirements.txt
Flask
Flask-Cors
selenium
webdriver-manager
python-dotenv
schedule
gunicorn
pytz
SQLAlchemy
psycopg2-binary
